refactor(evals): move the target matrix out of runner.ts
evalMatrix, reachable and Target only need the env and the provider
table, but living in runner.ts they dragged @repo/ai behind them. The
vitest config has to read the matrix to size the run, and it cannot pay
for the whole agent stack to do it.

Adds EVAL_CONCURRENCY here so the cap is a validated value like every
other one.
