+++ packages/evals/package.json
@@ -11,6 +11,7 @@
     "./package.json": "./package.json",
     "./report": "./src/report.ts",
     "./runner": "./src/runner.ts",
+    "./targets": "./src/targets.ts",
     "./trace": "./src/trace.ts"
   },
   "engines": {
